// functions/func-fofoprono-synchro/src/models/api-football/team.ts
export type Team =
  | "France"
  | "Germany"
  | "Portugal"
  | "Hungary"
  | "Spain"
  | "Sweden"
  | "Poland"
  | "Slovakia"
  | "Scotland"
  | "Czech Republic"
  | "Croatia"
  | "England"
  | "Austria"
  | "Netherlands"
  | "Ukraine"
  | "Serbia"
  | "Denmark"
  | "Finland"
  | "Belgium"
  | "Russia"
  | "Turkey"
  | "Wales"
  | "Italy"
  | "Switzerland"
  | "Georgia"
  | "Romania"
  | "Slovenia"
  | "Albania";