// functions/func-fofoprono-synchro/src/models/database/team.ts
export type Team =
  | "France"
  | "Allemagne"
  | "Portugal"
  | "Hongrie"
  | "Espagne"
  | "Suède"
  | "Pologne"
  | "Slovaquie"
  | "Écosse"
  | "République tchèque"
  | "Croatie"
  | "Angleterre"
  | "Autriche"
  | "Pays-Bas"
  | "Ukraine"
  | "Serbie"
  | "Danemark"
  | "Finlande"
  | "Belgique"
  | "Russie"
  | "Turquie"
  | "Pays de Galles"
  | "Italie"
  | "Suisse"
  | "Géorgie"
  | "Roumanie"
  | "Slovénie"
  | "Albanie";